+++ app/src/main/java/com/primeroeldev/mnemono/general/EntityInterface.kt
@@ -3,5 +3,5 @@ package com.primeroeldev.mnemono.general
 
 interface EntityInterface
 {
-    abstract var _id: Int?
+    abstract var _id: Long?
 }
